Selkeytetty index.html moduleilla ja script.js componenteilla. Myös server.js pilkottu pienempiin.
