+++ index.html
@@ -20,6 +20,7 @@
         <p>Welcome, <span id="userName"></span>!</p>
         <button id="logoutButton">Logout</button>
     </div>
+    
     <!-- Popup creation form -->
     <div id="popupForm" class="popup-form" style="display: none;">
         <a href="/">Home</a>
@@ -80,7 +81,7 @@
                 </div>
                 <input type="hidden" id="imageUrl" name="imageUrl">
             </div>
-                    <!-- Mahdollinen linkki eteenpäin -->
+            <!-- Mahdollinen linkki eteenpäin -->
             <div class="form-group">
                 <label for="linkUrl">Link URL (Valinnainen):</label>
                 <input type="url" id="linkUrl" name="linkUrl" placeholder="https://example.com">
@@ -112,17 +113,19 @@
             <button type="submit" id="createPopup">Create Popup</button>
         </form>
     </div>
+
     <div id="popupPreview" class="popup-preview" style="display: none;">
         <h3>Popup Preview</h3>
         <div id="previewPopup" class="preview-popup">
             <p id="previewContent">This is a preview of the popup content.</p>
         </div>
     </div>
-                    <!-- Popup list -->
-                    <div id="popupList" class="popup-list">
-                        <h3>Your Popups</h3>
-                        <ul id="popups"></ul>
-                        </div>
+    
+    <!-- Popup list -->
+    <div id="popupList" class="popup-list" style="display: none;">
+        <h3>Your Popups</h3>
+        <ul id="popups"></ul>
+    </div>
 
     <!-- kuvakirjasto -->
     <div id="imageLibrary" class="image-library" style="display: none;">
@@ -135,10 +138,10 @@
         <div class="image-grid" id="imageGrid">
           <!-- Kuvat ladataan tänne JavaScriptillä -->
         </div>
-      </div>
+    </div>
       
-      <!-- Lisää kuvavalitsin-dialogi -->
-      <div id="imagePickerDialog" class="modal" style="display: none;">
+    <!-- Lisää kuvavalitsin-dialogi -->
+    <div id="imagePickerDialog" class="modal" style="display: none;">
         <div class="modal-content">
           <span class="close-button" id="closeImagePicker">&times;</span>
           <h3>Valitse kuva</h3>
@@ -146,9 +149,10 @@
             <!-- Kuvat ladataan tänne JavaScriptillä -->
           </div>
         </div>
-      </div>
-      <!-- Kuvan tiedot -dialogi -->
-      <div id="imageDetailsDialog" class="modal" style="display: none;">
+    </div>
+    
+    <!-- Kuvan tiedot -dialogi -->
+    <div id="imageDetailsDialog" class="modal" style="display: none;">
         <div class="modal-content">
           <span class="close-button" id="closeImageDetails">&times;</span>
           <h3>Kuvan tiedot</h3>
@@ -167,7 +171,8 @@
             <ul id="detailsPopupList"></ul>
           </div>
         </div>
-      </div>
+    </div>
+
     <nav>
         <div id="adminMenu" style="display: none;">
             <h3>Admin Menu</h3>
@@ -179,10 +184,7 @@
         </div>
     </nav>
 
-
-
-
-     <!-- Edit popup form (hidden by default) -->
+    <!-- Edit popup form (hidden by default) -->
     <div id="editPopupForm" class="popup-form" style="display: none;">
         <h3>Edit Popup</h3>
         <form id="updatePopupForm">
@@ -278,7 +280,7 @@
                     <input type="datetime-local" id="editEndDate" name="endDate">
                 </div>
             </div>
-        <div class="preview" id="editPreview"></div>
+            <div class="preview" id="editPreview"></div>
 
             <div class="button-group">
                 <button type="submit">Update Popup</button>
@@ -286,7 +288,6 @@
             </div>
         </form>
     </div>
-
 
     <!-- Popup elements -->
     <button id="openPopup">Avaa popup</button>
@@ -297,7 +298,7 @@
         <button id="closePopup">Sulje</button>
     </div> 
 
-
-    <script src="script.js"></script>
+    <!-- Lataa JavaScript ES moduulit -->
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
